+++ config_test.go
@@ -246,7 +246,7 @@ Here are other value(loader)s:
 		assert.Equal(t,
 			"number has value[value:123] that is loaded by loader[map].\n\n",
 			config.Explain("number", konf.WithValueFormatter(
-				func(_ string, _ konf.Loader, value any) string {
+				func(_ konf.Loader, _ string, value any) string {
 					return fmt.Sprintf("value:%v", value)
 				},
 			)),